fixed typo and moved file
